Api/Dydx/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appnope==0.1.3
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bech32==1.2.0
bip-utils==2.9.1
bitarray==2.9.2
blspy==2.0.2
Brownie==0.5.1
cbor2==5.6.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coincurve==18.0.0
comm==0.2.1
crcmod==1.7
Cython==0.29.37
cytoolz==0.12.3
dateparser==1.2.0
debugpy==1.8.0
decorator==5.1.1
dydx-v4-python==0.0.4
ecdsa==0.18.0
ed25519-blake2b==1.4.1
eth-account==0.9.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.1
eth-typing==3.5.2
eth-utils==2.3.1
eth_abi==5.0.0
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
grpcio==1.60.1
grpcio-tools==1.60.1
hexbytes==0.3.1
idna==3.6
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
lru-dict==1.2.0
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
packaging==23.2
parsimonious==0.9.0
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-sr25519-bindings==0.2.0
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2024.1
pyunormalize==15.1.0
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.17.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
toolz==0.12.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.0
v4-client-py==1.0.1
v4-proto==2.0.0
wcwidth==0.2.13
web3==6.15.0
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
